fix(sonarqube): Limiter la duplication de code refers to https://trello.com/c/isf50e6H/2673
