+++ packages/api/server/services/userFavoriteShantytown/testHelpers.spec.ts
@@ -1,5 +1,6 @@
 import chai from 'chai';
 import ServiceError from '#server/errors/ServiceError';
+import { serialized as fakeUser } from '#test/utils/user';
 
 const { expect } = chai;
 
@@ -28,4 +29,34 @@ export async function expectPermissionDenied(
 export async function expectVoidReturn(fn: () => Promise<void>): Promise<void> {
     const result = await fn();
     expect(result).to.be.undefined;
+}
+
+/**
+ * Génère les tests de contrôle des permissions pour les rôles interdits
+ */
+export function testForbiddenRolesPermissions(
+    serviceFn: (user: any, shantytownId: number) => Promise<void>,
+    modelStub: sinon.SinonStub,
+    shantytownId: number,
+): void {
+    const forbiddenRoles = ['intervener', 'external_observator'] as const;
+
+    forbiddenRoles.forEach((roleId) => {
+        it(`rejette avec permission_denied si l'utilisateur a le rôle ${roleId}`, async () => {
+            const user = fakeUser({ role_id: roleId });
+            await expectPermissionDenied(() => serviceFn(user, shantytownId));
+        });
+
+        it(`n'appelle pas le modèle si l'utilisateur a le rôle ${roleId}`, async () => {
+            const user = fakeUser({ role_id: roleId });
+
+            try {
+                await serviceFn(user, shantytownId);
+            } catch {
+                // do nothing
+            }
+
+            expect(modelStub).to.not.have.been.called;
+        });
+    });
 }
